+++ crates/perl-corpus/src/api.rs
@@ -1,0 +1,33 @@
+pub use crate::cases::{
+    ComplexDataStructureCase, EdgeCase, EdgeCaseGenerator, complex_data_structure_cases,
+    edge_cases, find_complex_case, get_complex_data_structure_tests, sample_complex_case,
+};
+pub use crate::codegen::{
+    CodegenOptions, StatementKind, generate_perl_code, generate_perl_code_with_options,
+    generate_perl_code_with_seed, generate_perl_code_with_statements,
+};
+pub use crate::concepts::{ConceptRow, LoadedConcept, load_concept_registry};
+pub use crate::continue_redo::{
+    ContinueRedoCase, cases_by_tag as continue_redo_cases_by_tag, continue_redo_cases,
+    find_case as find_continue_redo_case, invalid_cases as invalid_continue_redo_cases,
+    valid_cases as valid_continue_redo_cases,
+};
+pub use crate::files::{
+    CORPUS_ROOT_ENV, CorpusFile, CorpusLayer, CorpusPaths, get_all_test_files, get_corpus_files,
+    get_corpus_files_from, get_fuzz_files, get_test_files,
+};
+pub use crate::format_statements::{
+    FormatStatementCase, FormatStatementGenerator, find_format_case, format_statement_cases,
+};
+pub use crate::glob_expressions::{
+    GlobExpressionCase, GlobExpressionGenerator, find_glob_case, glob_expression_cases,
+};
+pub use crate::gold::*;
+pub use crate::inventory::*;
+pub use crate::loading::{parse_dir, parse_file};
+pub use crate::metadata::{Section, find_by_flag, find_by_tag};
+pub use crate::sidecar::*;
+pub use crate::tie_interface::{
+    TieInterfaceCase, find_tie_case, tie_cases_by_tag, tie_cases_by_tags_all,
+    tie_cases_by_tags_any, tie_interface_cases,
+};
