+++ crates/perl-corpus/src/metadata/flags.rs
@@ -1,0 +1,3 @@
+pub fn normalize_flags(raw: &str) -> Vec<String> {
+    raw.replace(',', " ").split_whitespace().map(str::to_string).collect()
+}
