+++ crates/perl-corpus/src/metadata/tags.rs
@@ -1,0 +1,3 @@
+pub fn normalize_tags(raw: &str) -> Vec<String> {
+    raw.replace(',', " ").split_whitespace().map(str::to_lowercase).collect()
+}
